Reorganize application components
- Move application detail drawer components to application-detail-drawer
- Move application form components to application-form
- Move application identity form components to application-identity-form
- Move application analysis status components to applications-table
- Move application assessment status components to applications-table
- Move application review status components to applications-table
- Move import applications form components to import-applications-form
