+++ client/src/app/pages/applications/application-detail-drawer/fact-detail-modal/fact-code-snip-viewer.tsx
