+++ client/src/app/pages/applications/application-detail-drawer/fact-detail-modal/fact-detail-modal.tsx
