+++ client/src/app/pages/applications/application-detail-drawer/fact-detail-modal/index.ts
