+++ client/src/app/pages/applications/application-form/__tests__/application-form.test.tsx
