+++ client/src/app/pages/applications/applications-table/components/application-assessment-status.tsx
