--- client/src/app/pages/applications/components/application-assessment-status/index.ts
@@ -1,1 +1,0 @@
-export { ApplicationAssessmentStatus } from "./application-assessment-status";
