--- client/src/app/pages/applications/components/import-applications-form/index.ts
@@ -1,1 +1,0 @@
-export { ImportApplicationsForm } from "./import-applications-form";
